+++ modyn/config/schema/pipeline.py
@@ -4,12 +4,13 @@ from functools import cached_property
 from pathlib import Path
 from typing import Annotated, Any, Dict, List, Literal, Optional, Union
 
-from modyn.supervisor.internal.eval_strategies import OffsetEvalStrategy
-from modyn.utils import validate_timestr
-from modyn.utils.utils import SECONDS_PER_UNIT
 from pydantic import BaseModel as PydanticBaseModel
 from pydantic import Field, NonNegativeInt, field_validator, model_validator
 from typing_extensions import Self
+
+from modyn.supervisor.internal.eval_strategies import OffsetEvalStrategy
+from modyn.utils import validate_timestr
+from modyn.utils.utils import SECONDS_PER_UNIT
 
 # ----------------------------------------------------- PIPELINE ----------------------------------------------------- #
 
@@ -429,9 +430,6 @@ class TimeTriggerConfig(BaseModel):
         description="Interval length for the trigger as an integer followed by a time unit: s, m, h, d, w, y",
         pattern=rf"^\d+{_REGEX_TIME_UNIT}$",
     )
-    sample_size: int | None = Field(None, description="The number of samples used for the metric calculation.", ge=1)
-    metric: str = Field("model", description="The metric used for drift detection.")
-    metric_config: dict[str, Any] = Field(default_factory=dict, description="Configuration for the evidently metric.")
 
     @cached_property
     def every_seconds(self) -> int:
